feat: add missing gender property on person

+++ core/network/src/main/kotlin/com/divinelink/core/network/media/model/credits/SeriesCastApi.kt
@@ -8,7 +8,7 @@ data class SeriesCastApi(
   val id: Long,
   val name: String,
   val adult: Boolean,
-  val gender: Int,
+  val gender: Long,
   val popularity: Double,
   val roles: List<RolesApi>,
   val order: Int,
